requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.11.2
django-tenant-schemas==1.11.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-storages==1.14.2
boto3==1.29.7
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0
stripe==7.6.0
PyJWT==2.8.0
django-filter==23.4
django-extensions==3.2.3
django-debug-toolbar==4.2.0
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2